% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsws-package.R
\name{Client}
\alias{Client}
\title{R6 class for OBS WebSocket client}
\value{
\code{new(Client)} returns a new OBSClient object.

\code{client$pluck()} returns JSON strings representing recieved messages.

\code{client$emit()} returns corresponding request ID sent to OBS.

\code{client$current_state()} returns current state of the client.

\code{client$ready_state()} returns current WebSocket ready state.
}
\description{
An R6 class for OBS WebSocket server client.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OBSClient-new}{\code{Client$new()}}
\item \href{#method-OBSClient-connect}{\code{Client$connect()}}
\item \href{#method-OBSClient-disconnect}{\code{Client$disconnect()}}
\item \href{#method-OBSClient-reset}{\code{Client$reset()}}
\item \href{#method-OBSClient-pluck}{\code{Client$pluck()}}
\item \href{#method-OBSClient-emit}{\code{Client$emit()}}
\item \href{#method-OBSClient-reidentify}{\code{Client$reidentify()}}
\item \href{#method-OBSClient-current_state}{\code{Client$current_state()}}
\item \href{#method-OBSClient-ready_state}{\code{Client$ready_state()}}
\item \href{#method-OBSClient-clone}{\code{Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-new}{}}}
\subsection{Method \code{new()}}{
Creates a new OBSClient object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(url = "ws://localhost:4455", id = rand_id(1))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL of an OBS WebSocket server.}

\item{\code{id}}{Client ID (used as prefix for request IDs).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-connect"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-connect}{}}}
\subsection{Method \code{connect()}}{
Connects to an OBS WebSocket server.
Note that this method behaves synchronously (blocks the R session)
so as to wait for the identified message from the server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$connect(password, max_retries = 10, interval = 1, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{password}}{Password for authentication.}

\item{\code{max_retries}}{Maximum number of retries.}

\item{\code{interval}}{Interval between retries.}

\item{\code{verbose}}{Whether to print messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnects from the OBS WebSocket server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$disconnect(reset = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reset}}{Whether to reset message queue after disconnect.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-reset"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-reset}{}}}
\subsection{Method \code{reset()}}{
Resets message queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-pluck"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-pluck}{}}}
\subsection{Method \code{pluck()}}{
Gets messages from queue as a character vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$pluck(only, reset = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only}}{Indices of messages to pluck.
If missing, all messages are plucked.}

\item{\code{reset}}{Whether to reset queue after pluck.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-emit"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-emit}{}}}
\subsection{Method \code{emit()}}{
Sends a request to OBS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$emit(type, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string representing request type.}

\item{\code{data}}{A list of request data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-reidentify"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-reidentify}{}}}
\subsection{Method \code{reidentify()}}{
Reidentifies with event subscriptions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$reidentify(subscriptions = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subscriptions}}{An integer scalar
representing event subscriptions bitmask.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-current_state"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-current_state}{}}}
\subsection{Method \code{current_state()}}{
Gets current client state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$current_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-ready_state"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-ready_state}{}}}
\subsection{Method \code{ready_state()}}{
Gets current WebSocket ready state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$ready_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OBSClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OBSClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
