% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EventSub}
\alias{EventSub}
\title{OBS WebSocket event subscription bitmasks}
\format{
An object of class \code{list} of length 17.
}
\source{
\url{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.json}
}
\usage{
EventSub
}
\description{
A named list of integer bitmask values for event subscriptions
used by obs-websocket (version 5.x).
}
\seealso{
Other obsws-data: 
\code{\link{ReqType}},
\code{\link{obs_enums}},
\code{\link{obs_events}},
\code{\link{obs_requests}}
}
\concept{obsws-data}
\keyword{datasets}
