% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_enums}
\alias{obs_enums}
\title{Enums from the OBS WebSocket protocol}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\source{
\url{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.json}
}
\usage{
obs_enums
}
\description{
See \href{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.md#enums}{obs-websocket/docs/generated/protocol.md}
}
\seealso{
Other obsws-data: 
\code{\link{ReqType}},
\code{\link{obs_events}},
\code{\link{obs_requests}}
}
\concept{obsws-data}
\keyword{datasets}
