% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_events}
\alias{obs_events}
\title{Events from the OBS WebSocket protocol}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57 rows and 9 columns.
}
\source{
\url{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.json}
}
\usage{
obs_events
}
\description{
See \href{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.md#events}{obs-websocket/docs/generated/protocol.md}
}
\seealso{
Other obsws-data: 
\code{\link{ReqType}},
\code{\link{obs_enums}},
\code{\link{obs_requests}}
}
\concept{obsws-data}
\keyword{datasets}
