% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_requests}
\alias{obs_requests}
\title{Requests from the OBS WebSocket protocol}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 146 rows and 9 columns.
}
\source{
\url{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.json}
}
\usage{
obs_requests
}
\description{
See \href{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.md#requests}{obs-websocket/docs/generated/protocol.md}
}
\seealso{
Other obsws-data: 
\code{\link{ReqType}},
\code{\link{obs_enums}},
\code{\link{obs_events}}
}
\concept{obsws-data}
\keyword{datasets}
