% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsws-package.R
\docType{package}
\name{obsws-package}
\alias{obsws}
\alias{obsws-package}
\title{obsws: Minimal Client for 'OBS' WebSocket}
\description{
A minimal R client for the 'OBS' WebSocket protocol ('obs-websocket' 5.x, \url{https://github.com/obsproject/obs-websocket}). This package provides low-level utilities to send requests and receive messages as raw JSON strings via WebSocket connections. It is designed for interactive use and does not attempt to abstract or validate 'OBS' request semantics.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
