% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser}
\alias{parser}
\alias{parse_message}
\alias{parse_op}
\alias{parse_data}
\alias{parse_event_type}
\title{Parse messages from the OBS WebSocket server}
\usage{
parse_message(jsonl, only)

parse_op(jsonl, only)

parse_data(jsonl, only, query = "/d")

parse_event_type(jsonl, only)
}
\arguments{
\item{jsonl}{A character vector of JSON messages.}

\item{only}{Indices of messages to parse.
If missing, all messages are parsed.}

\item{query}{Query to be passed to \code{\link[RcppSimdJson:fparse]{RcppSimdJson::fparse()}}.
Defaults to \code{"/d"}.}
}
\value{
\itemize{
\item For \code{parse_message()}, a list of parsed messages.
\item For \code{parse_op()}, an integer vector of parsed '/op' fields.
\item For \code{parse_data()}, a list of parsed fields.
\item For \code{parse_request_id()}, a character vector of parsed '/d/requestId' fields.
\item For \code{parse_event_type()}, a character vector of parsed '/d/eventType' fields.
}
}
\description{
These functions parse JSON messages from the OBS WebSocket server.
They are wrappers around \code{\link[RcppSimdJson:fparse]{RcppSimdJson::fparse()}}.
}
