% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rand_id}
\alias{rand_id}
\title{Generate random IDs}
\usage{
rand_id(n, length = 12)
}
\arguments{
\item{n}{Number of IDs to generate}

\item{length}{Length of each ID}
}
\value{
A character vector
}
\description{
Generate random IDs
}
\keyword{internal}
