% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReqType}
\alias{ReqType}
\title{All request types from the OBS WebSocket protocol}
\format{
An object of class \code{list} of length 146.
}
\source{
\url{https://github.com/obsproject/obs-websocket/blob/master/docs/generated/protocol.json}
}
\usage{
ReqType
}
\description{
All request types from the OBS WebSocket protocol
}
\seealso{
Other obsws-data: 
\code{\link{obs_enums}},
\code{\link{obs_events}},
\code{\link{obs_requests}}
}
\concept{obsws-data}
\keyword{datasets}
