% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiters.R
\name{wait_for_messages}
\alias{wait_for_messages}
\title{Wait for messages}
\usage{
wait_for_messages(timeout = 1)
}
\arguments{
\item{timeout}{Maximum time to wait for messages.}
}
\value{
Called for side effects.
}
\description{
Waits for messages from the OBS websocket server
until a timeout is reached.

This is useful for forcing to update the message queue of a client.
Since the message queue is updated asynchronously
within an event loop of the \link[later:later]{later::later} package,
you need to update the queue manually where the R session doesn't get idle.
}
