% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiters.R
\name{waiter_for_response}
\alias{waiter_for_response}
\title{Generate waiter for RequestResponse messages}
\usage{
waiter_for_response(max_retries = 10, interval = 1, error_on_timeout = TRUE)
}
\arguments{
\item{max_retries}{Maximum number of retries.}

\item{interval}{Interval between retries.}

\item{error_on_timeout}{Whether to throw an error if no response is received.}
}
\value{
A function that takes a \code{Client} and request ID as arguments
and returns the response message as a JSON string from the queue.
}
\description{
Generate waiter for RequestResponse messages
}
