% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsws-package.R
\docType{package}
\name{obsws-package}
\alias{obsws}
\alias{obsws-package}
\title{obsws: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
